server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
jinja2>=3.1.2
pillow>=10.0.0
pytest>=7.0.0
httpx>=0.25.0
pytest-asyncio>=1.0.0
